Fix race condition when setting up handlers.
onHTTPRequest and onWebsocketRequest handlers was initialized after the
client was added to epoll. This might cause a raceconditon where we try
to parse a client request before it has an assigned handler. When this
happens a std::bad_function_call is thrown. Fixed by setting up the
handlers before adding the client to our epoll set.